Merged messed up trunk with DomainManager branch.
DomainManager shifted to private static within the FieldML class. FieldML class methods will eventually be native methods living in a library, so passing objects as parameters is a no-go.

If a static DomainManager is too objectionable, passing some kind of 'session id' across, and resolving that into a session-specific manager would work.