The FieldML implementation is now an object explicitly wrapped in an interface. The FieldML interface is now a simple list of methods comprising the API, which implementation details shifted to the FieldMLJava class.
Added the FieldParameters class, analogous to the cache concept from cmgui.

Real and Index valued fields can now be defined and evaluated over discrete domains.

The Value class is now a union-analog. This removes the need for fields and components to use casts and instanceofs. Because the definition of fields and their components is static (or at least write-once), type-checking only needs to be done at instantion time (i.e. when parsing the XML).
Note: A lot of this sort of checking is not actually being done at the moment.