To ensure that field-type discrete domains remain valid (i.e. member fields all have the same signature), the API needs to have a concept of 'finished' field and domain declarations. The user will only have access to the ids of 'finished' fields and domains, so field-type domains can be checked at their declaration, and because their member fields are now immutable, the domain will remain valid.
The will mean that we can no longer define all domains first, and then all fields. Field-type domains will need a list of field ids for their definition, meaning that as well as fields that depend on domains, some domains will in turn depend on fields. This will require the XML parser to construct a dependancy tree before instantiating fields and domains, and it precludes domain-field circular dependancies that involve field-type domains.